--- app/src/main/java/com/lateinit/rightweight/data/datasource/BaseDataSource.kt
@@ -1,4 +1,0 @@
-package com.lateinit.rightweight.data.datasource
-
-class BaseDataSource {
-}
