+++ app/src/main/java/com/lateinit/rightweight/data/datasource/local/SharedRoutineLocalDataSource.kt
@@ -1,4 +1,4 @@
-package com.lateinit.rightweight.data.datasource
+package com.lateinit.rightweight.data.datasource.local
 
 import com.lateinit.rightweight.data.database.entity.SharedRoutineDay
 import com.lateinit.rightweight.data.database.entity.SharedRoutineExercise
