+++ app/src/main/java/com/lateinit/rightweight/data/datasource/remote/LoginDataSource.kt
@@ -1,4 +1,4 @@
-package com.lateinit.rightweight.data.datasource
+package com.lateinit.rightweight.data.datasource.remote
 
 import com.lateinit.rightweight.data.LoginResponse
