+++ app/src/main/java/com/lateinit/rightweight/data/repository/impl/UserRepositoryImpl.kt
@@ -1,8 +1,8 @@
 package com.lateinit.rightweight.data.repository.impl
 
 import com.lateinit.rightweight.data.database.intermediate.RoutineWithDays
-import com.lateinit.rightweight.data.datasource.UserLocalDataSource
-import com.lateinit.rightweight.data.datasource.UserRemoteDataSource
+import com.lateinit.rightweight.data.datasource.local.UserLocalDataSource
+import com.lateinit.rightweight.data.datasource.remote.UserRemoteDataSource
 import com.lateinit.rightweight.data.model.User
 import com.lateinit.rightweight.data.model.WriteModelData
 import com.lateinit.rightweight.data.repository.UserRepository
