+++ app/src/main/java/com/lateinit/rightweight/di/DataSourceModule.kt
@@ -10,6 +10,19 @@ import com.lateinit.rightweight.data.database.dao.RoutineDao
 import com.lateinit.rightweight.data.database.dao.SharedRoutineDao
 import com.lateinit.rightweight.data.database.dao.UserDao
 import com.lateinit.rightweight.data.datasource.*
+import com.lateinit.rightweight.data.datasource.local.impl.HistoryLocalLocalDataSourceImpl
+import com.lateinit.rightweight.data.datasource.local.RoutineLocalDataSource
+import com.lateinit.rightweight.data.datasource.local.impl.RoutineLocalDataSourceImpl
+import com.lateinit.rightweight.data.datasource.local.UserLocalDataSource
+import com.lateinit.rightweight.data.datasource.local.impl.UserLocalDataSourceImpl
+import com.lateinit.rightweight.data.datasource.remote.LoginDataSource
+import com.lateinit.rightweight.data.datasource.remote.impl.LoginDataSourceImpl
+import com.lateinit.rightweight.data.datasource.local.SharedRoutineLocalDataSource
+import com.lateinit.rightweight.data.datasource.local.impl.SharedRoutineLocalDataSourceImpl
+import com.lateinit.rightweight.data.datasource.remote.SharedRoutineRemoteDataSource
+import com.lateinit.rightweight.data.datasource.remote.impl.SharedRoutineRemoteDataSourceImpl
+import com.lateinit.rightweight.data.datasource.remote.UserRemoteDataSource
+import com.lateinit.rightweight.data.datasource.remote.impl.UserRemoteDataSourceImpl
 import dagger.Module
 import dagger.Provides
 import dagger.hilt.InstallIn
